Fixed the error causing the transaction page not to show.

+++ lib/utils/Transaction.dart
@@ -7,6 +7,7 @@ class Transaction{
   Transaction({this.amount, this.description, this.transactionDate});
 
   static Transaction fromJson(Map<String, dynamic> json){
-    return Transaction(amount: json['amount'], description: json['description'], transactionDate: json['transactionDate']);
+
+    return Transaction(amount: json['amount'].toDouble(), description: json['description'], transactionDate: json['transactionDate']);
   }
 }
